% Generated by roxygen2 (4.0.1): do not edit by hand
\name{near_match}
\alias{near_match}
\title{Near match}
\usage{
near_match(a_word, normalize_ending = FALSE)
}
\arguments{
\item{a_word}{(character) A character string. Required.}

\item{normalize_ending}{(logical) Whether to normalize the ending of the string. Default: FALSE}
}
\description{
Near match
}
\examples{
near_match(a_word='Helianthus   annuus')
}

