% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize_author}
\alias{normalize_author}
\title{Normalize an author}
\usage{
normalize_author(string)
}
\arguments{
\item{string}{A string of author(s)}
}
\description{
Normalize an author
}
\examples{
normalize_author(string='Gunther, Brunkal')
}

