% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize_ending}
\alias{normalize_ending}
\title{Noramlize endings.}
\usage{
normalize_ending(a_word)
}
\arguments{
\item{a_word}{(character) A character string. Required.}
}
\description{
Noramlize endings.
}
\examples{
normalize_ending(a_word='HELIANTHIS')
normalize_ending(a_word='TIM')
normalize_ending(a_word='RAXTAS')
}

