% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize_word}
\alias{normalize_word}
\title{Normalize a word}
\usage{
normalize_word(word)
}
\arguments{
\item{word}{A word}
}
\description{
Normalize a word
}
\examples{
normalize_word(word='Helianthus  ')
normalize_word(word='G9-Helianthus  ')
}

