% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phonetize}
\alias{phonetize}
\title{Phonetize}
\usage{
phonetize(a_word, normalize_ending = FALSE)
}
\arguments{
\item{a_word}{(character) A character string. Required.}

\item{normalize_ending}{(logical) Whether to normalize the ending of the string. Default: FALSE}
}
\description{
Phonetize
}

