% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tr}
\alias{tr}
\title{Replace x with y in a string}
\usage{
tr(str, src, repl)
}
\arguments{
\item{str}{Target string}

\item{src}{Characters to replace}

\item{repl}{Replacement characters}
}
\description{
Replace x with y in a string
}
\examples{
tr(str='HELIANTHUS ANNUS', src='EOUYKZ', repl='IAIICS')
}

